\name{asthma}
\docType{data}
\alias{asthma}
\title{Asthma recurrence in children}
\description{ 
  This data set gives the start and stop times of recurrent asthma events in children. It also provides a subject ID, treatment indicator, censoring indicator, number of events per subject and a first event indicator.
}
\format{A data frame with 1037 rows and 7 columns. See asthma.txt header for details.}
\source{http://www.blackwellpublishing.com/rss/}
\references{Duchateau et al. \emph{JRSSC} 2003. Volume 52, 355--363.}
\keyword{datasets}
