% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condMoments.R
\name{K1.norm}
\alias{K1.norm}
\title{K1.norm}
\usage{
K1.norm(w)
}
\arguments{
\item{w}{real value}
}
\value{
conditional expectation
}
\description{
\code{E(Y|Y>w)} where Y is normal
}
\author{
David Clement
}
