% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condMoments.R
\name{K2.exp}
\alias{K2.exp}
\title{K2.exp}
\usage{
K2.exp(w)
}
\arguments{
\item{w}{real value}
}
\value{
conditional expectation
}
\description{
\code{E(Y^2|Y>w)} where Y is exponential dist with mean 0
and variance 1
}
\author{
David Clement
}
