% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condMoments.R
\name{K2.norm}
\alias{K2.norm}
\title{K2.norm}
\usage{
K2.norm(w)
}
\arguments{
\item{w}{real value}
}
\value{
conditional expectation
}
\description{
\code{E(Y^2|Y>w)} where Y is normal
}
\author{
David Clement
}
