% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condMoments.R
\name{K2.t3}
\alias{K2.t3}
\title{K2.t3}
\usage{
K2.t3(w)
}
\arguments{
\item{w}{real value}
}
\value{
conditional expectation
}
\description{
\code{E(Y^2|Y>w)} where Y is t dist with 3 df
}
\author{
David Clement
}
